+++ clease/settings/settings.py
@@ -109,6 +109,7 @@ class ClusterExpansionSettings:
         include_background_atoms: bool = False,
         basis_func_type="polynomial",
     ) -> None:
+        # pylint: disable=too-many-arguments
 
         self._include_background_atoms = include_background_atoms
         self._cluster_mng = None